Minimal sdk
