+++ cardano-sdk-playground/cardano-embedded-sdk/src/coin.rs
@@ -12,8 +12,7 @@ pub const MAX_COIN: u64 = 45_000_000_000__000_000;
 
 /// error type relating to `Coin` operations
 ///
-#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Copy, Serialize, Deserialize)]
-// #[cfg_attr(feature = "generic-serialization", derive(Serialize, Deserialize))]
+#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Copy)]
 pub enum Error {
     /// means that the given value was out of bound
     ///
@@ -53,8 +52,7 @@ pub enum CoinDiff {
 
 // TODO: add custom implementation of `serde::de::Deserialize` so we can check the
 // upper bound of the `Coin`.
-#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Copy, Serialize, Deserialize)]
-// #[cfg_attr(feature = "generic-serialization", derive(Serialize, Deserialize))]
+#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Copy)]
 pub struct Coin(u64);
 impl Coin {
     /// create a coin of value `0`.
@@ -127,7 +125,7 @@ impl ::core::str::FromStr for Coin {
         Coin::new(v)
     }
 }
-// TODO: rewrite with minicbor
+// TODO: cbor
 // impl cbor_event::se::Serialize for Coin {
 //     fn serialize<'se, W: ::std::io::Write>(
 //         &self,
