+++ cardano-sdk-playground/cardano-embedded-sdk/src/hash.rs
@@ -8,17 +8,13 @@ use core::{
     str::FromStr,
 };
 
+use crate::util::{hex, try_from_slice::TryFromSlice};
+
 use cryptoxide::blake2b::Blake2b;
 use cryptoxide::digest::Digest;
 use cryptoxide::sha3::Sha3;
 
-use util::{hex, try_from_slice::TryFromSlice};
-
-// #[cfg(feature = "generic-serialization")]
-use serde;
-
-#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Copy, Serialize, Deserialize)]
-// #[cfg_attr(feature = "generic-serialization", derive(Serialize, Deserialize))]
+#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Clone, Copy)]
 pub enum Error {
     InvalidHashSize(usize, usize),
     HexadecimalError(hex::Error),
@@ -41,7 +37,7 @@ impl From<hex::Error> for Error {
     }
 }
 impl ::core::error::Error for Error {
-    fn cause(&self) -> Option<&dyn core::error::Error> {
+    fn cause(&self) -> Option<&dyn (::core::error::Error)> {
         match self {
             Error::HexadecimalError(ref err) => Some(err),
             Error::InvalidHashSize(_, _) => None,
@@ -125,7 +121,7 @@ macro_rules! define_hash_object {
                 Self::try_from_slice(&bytes)
             }
         }
-        // TODO: rewrite with minicbor
+        // TODO: cbor
         // impl cbor_event::de::Deserialize for $hash_ty {
         //     fn deserialize<R: BufRead>(reader: &mut Deserializer<R>) -> cbor_event::Result<Self> {
         //         let bytes = reader.bytes()?;
@@ -149,69 +145,69 @@ macro_rules! define_hash_object {
         //         serializer.write_bytes(self.as_ref())
         //     }
         // }
-
+        //
         // #[cfg(feature = "generic-serialization")]
-        impl serde::Serialize for $hash_ty {
-            #[inline]
-            fn serialize<S>(&self, serializer: S) -> result::Result<S::Ok, S::Error>
-            where
-                S: serde::Serializer,
-            {
-                if serializer.is_human_readable() {
-                    serializer.serialize_str(&hex::encode(self.as_ref()))
-                } else {
-                    serializer.serialize_bytes(&self.as_ref())
-                }
-            }
-        }
+        // impl serde::Serialize for $hash_ty {
+        //     #[inline]
+        //     fn serialize<S>(&self, serializer: S) -> result::Result<S::Ok, S::Error>
+        //     where
+        //         S: serde::Serializer,
+        //     {
+        //         if serializer.is_human_readable() {
+        //             serializer.serialize_str(&hex::encode(self.as_ref()))
+        //         } else {
+        //             serializer.serialize_bytes(&self.as_ref())
+        //         }
+        //     }
+        // }
         // #[cfg(feature = "generic-serialization")]
-        impl<'de> serde::Deserialize<'de> for $hash_ty {
-            fn deserialize<D>(deserializer: D) -> result::Result<Self, D::Error>
-            where
-                D: serde::Deserializer<'de>,
-            {
-                struct HashVisitor;
-                impl<'de> serde::de::Visitor<'de> for HashVisitor {
-                    type Value = $hash_ty;
-
-                    fn expecting(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
-                        write!(fmt, "Expecting a Blake2b_256 hash (`Hash`)")
-                    }
-
-                    fn visit_str<'a, E>(self, v: &'a str) -> result::Result<Self::Value, E>
-                    where
-                        E: serde::de::Error,
-                    {
-                        match Self::Value::from_str(&v) {
-                            Err(Error::HexadecimalError(err)) => Err(E::custom(format!("{}", err))),
-                            Err(Error::InvalidHashSize(sz, _)) => {
-                                Err(E::invalid_length(sz, &"32 bytes"))
-                            }
-                            Ok(h) => Ok(h),
-                        }
-                    }
-
-                    fn visit_bytes<'a, E>(self, v: &'a [u8]) -> result::Result<Self::Value, E>
-                    where
-                        E: serde::de::Error,
-                    {
-                        match Self::Value::try_from_slice(v) {
-                            Err(Error::InvalidHashSize(sz, _)) => {
-                                Err(E::invalid_length(sz, &"32 bytes"))
-                            }
-                            Err(err) => panic!("unexpected error: {}", err),
-                            Ok(h) => Ok(h),
-                        }
-                    }
-                }
-
-                if deserializer.is_human_readable() {
-                    deserializer.deserialize_str(HashVisitor)
-                } else {
-                    deserializer.deserialize_bytes(HashVisitor)
-                }
-            }
-        }
+        // impl<'de> serde::Deserialize<'de> for $hash_ty {
+        //     fn deserialize<D>(deserializer: D) -> result::Result<Self, D::Error>
+        //     where
+        //         D: serde::Deserializer<'de>,
+        //     {
+        //         struct HashVisitor;
+        //         impl<'de> serde::de::Visitor<'de> for HashVisitor {
+        //             type Value = $hash_ty;
+        //
+        //             fn expecting(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
+        //                 write!(fmt, "Expecting a Blake2b_256 hash (`Hash`)")
+        //             }
+        //
+        //             fn visit_str<'a, E>(self, v: &'a str) -> result::Result<Self::Value, E>
+        //             where
+        //                 E: serde::de::Error,
+        //             {
+        //                 match Self::Value::from_str(&v) {
+        //                     Err(Error::HexadecimalError(err)) => Err(E::custom(format!("{}", err))),
+        //                     Err(Error::InvalidHashSize(sz, _)) => {
+        //                         Err(E::invalid_length(sz, &"32 bytes"))
+        //                     }
+        //                     Ok(h) => Ok(h),
+        //                 }
+        //             }
+        //
+        //             fn visit_bytes<'a, E>(self, v: &'a [u8]) -> result::Result<Self::Value, E>
+        //             where
+        //                 E: serde::de::Error,
+        //             {
+        //                 match Self::Value::try_from_slice(v) {
+        //                     Err(Error::InvalidHashSize(sz, _)) => {
+        //                         Err(E::invalid_length(sz, &"32 bytes"))
+        //                     }
+        //                     Err(err) => panic!("unexpected error: {}", err),
+        //                     Ok(h) => Ok(h),
+        //                 }
+        //             }
+        //         }
+        //
+        //         if deserializer.is_human_readable() {
+        //             deserializer.deserialize_str(HashVisitor)
+        //         } else {
+        //             deserializer.deserialize_bytes(HashVisitor)
+        //         }
+        //     }
+        // }
     };
 }
