+++ cardano-sdk-playground/cardano-embedded-sdk/src/lib.rs
@@ -4,10 +4,16 @@
 extern crate alloc;
 
 pub mod address;
+pub mod coin;
 pub mod config;
+pub mod hash;
 pub mod hdpayload;
 pub mod hdwallet;
+pub mod redeem;
+pub mod tx;
 pub mod util;
 
 pub mod bip;
 pub mod wallet;
+
+pub mod tags;
