+++ cardano-sdk-playground/cardano-embedded-sdk/src/util/base58.rs
@@ -19,8 +19,7 @@ use alloc::{
 
 pub const ALPHABET: &'static str = "123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz";
 
-#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Copy, Clone, Serialize, Deserialize)]
-// #[cfg_attr(feature = "generic-serialization", derive(Serialize, Deserialize))]
+#[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Copy, Clone)]
 pub enum Error {
     /// error when a given character is not part of the supported
     /// base58 `ALPHABET`. Contains the index of the faulty byte.
