+++ cardano-sdk-playground/cardano-embedded-sdk/src/util/mod.rs
@@ -1,3 +1,5 @@
+pub mod base58;
 pub mod bits;
 pub mod hex;
 pub mod securemem;
+pub mod try_from_slice;
