+++ cardano-sdk-playground/cardano-embedded-sdk/src/util/try_from_slice.rs
