+++ cardano-sdk-playground/cardano-origin/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cardano-tmp"
+name = "cardano-origin"
 version = "0.1.1"
 authors = ["Vincent Hanquez <vincent.hanquez@iohk.io>", "Nicolas Di Prima <nicolas.diprima@iohk.io>"]
 license = "MIT OR Apache-2.0"
@@ -11,15 +11,21 @@ The Cardano library. This includes full support for the Cardano blockchain and w
 """
 keywords = [ "Cardano", "Wallet", "Crypto" ]
 
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+name = "cardano-scenario"
+path = "src/main.rs"
+
 [build-dependencies]
 
 [dependencies]
 cryptoxide = "0.1"
 cbor_event = "^2.1.1"
 
-ufmt = "0.2.0"
-serde = { version = "1.0", default-features = false }
-serde_derive = { version = "1.0", default-features = false }
+serde = { version = "1.0", optional = true }
+serde_derive = { version = "1.0", optional = true }
 
 [dev-dependencies]
 lazy_static = "1.2"
@@ -28,3 +34,8 @@ serde_json = "1.0"
 unicode-normalization = "0.1"
 quickcheck = "0.8"
 base64 = "0.9"
+
+[features]
+default = []
+with-bench = []
+generic-serialization = ["serde", "serde_derive"]
