+++ cardano-sdk-playground/cardano-origin/test-vectors/network-magic.cbor
