--- cardano-sdk-playground/cardano-tmp/src/bip/mod.rs
@@ -1,8 +1,0 @@
-//! BitCoin improvements proposal implementations
-//!
-//! This intends to provide convenient tooling for user of the cardano-tmp
-//! library for managing wallet, private keys and addresses.
-//!
-
-pub mod bip39;
-// pub mod bip44;
