--- cardano-sdk-playground/cardano-tmp/src/util/mod.rs
@@ -1,8 +1,0 @@
-pub mod base58;
-pub mod bits;
-pub mod hex;
-pub mod securemem;
-pub mod try_from_slice;
-
-#[cfg(test)]
-pub mod arbitrary;
