--- cardano-sdk-playground/cardano-tmp/src/wallet/mod.rs
@@ -1,4 +1,0 @@
-// pub mod bip44;
-pub mod keygen;
-pub mod rindex;
-pub mod scheme;
